Move ts extension tests to jest extension
