+++ extensions/jest/src/test/java/io/doov/ts/ast/test/TypeScriptIterableTest.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (C) by Courtanet, All Rights Reserved.
  */
-package io.doov.ts.ast;
+package io.doov.ts.ast.test;
 
 import static io.doov.assertions.ts.Assertions.assertThat;
 import static io.doov.core.dsl.meta.i18n.ResourceBundleProvider.BUNDLE;
@@ -24,6 +24,7 @@ import io.doov.core.dsl.field.types.IterableFieldInfo;
 import io.doov.core.dsl.lang.MappingRule;
 import io.doov.core.dsl.meta.Metadata;
 import io.doov.core.dsl.runtime.GenericModel;
+import io.doov.ts.ast.AstTSRenderer;
 import io.doov.ts.ast.writer.DefaultTypeScriptWriter;
 import io.doov.ts.ast.writer.TypeScriptWriter;
 import io.doov.tsparser.TypeScriptParser;
