+++ extensions/jest/src/test/java/io/doov/ts/ast/test/TypeScriptMatchAllTest.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.doov.ts.ast;
+package io.doov.ts.ast.test;
 
-import static io.doov.assertions.ts.Assertions.assertParenthesis;
 import static io.doov.assertions.ts.Assertions.assertThat;
 import static io.doov.core.dsl.DOOV.alwaysFalse;
 import static io.doov.core.dsl.DOOV.alwaysTrue;
@@ -24,19 +23,19 @@ import static io.doov.core.dsl.DOOV.when;
 import static io.doov.core.dsl.meta.i18n.ResourceBundleProvider.BUNDLE;
 import static io.doov.tsparser.util.TypeScriptParserFactory.parseUsing;
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static java.util.Arrays.asList;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.util.List;
+import java.time.LocalDate;
 import java.util.Locale;
 import java.util.function.Function;
 
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.antlr.v4.runtime.tree.ParseTreeWalker;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Test;
 
 import io.doov.assertions.ts.TypeScriptAssertionContext;
 import io.doov.core.dsl.field.types.*;
@@ -44,41 +43,64 @@ import io.doov.core.dsl.lang.Result;
 import io.doov.core.dsl.lang.StepCondition;
 import io.doov.core.dsl.meta.Metadata;
 import io.doov.core.dsl.runtime.GenericModel;
+import io.doov.core.dsl.time.LocalDateSuppliers;
+import io.doov.ts.ast.AstTSRenderer;
 import io.doov.ts.ast.writer.DefaultTypeScriptWriter;
 import io.doov.ts.ast.writer.TypeScriptWriter;
 import io.doov.tsparser.TypeScriptParser;
 
-class TypeScriptCombinedTest {
+class TypeScriptMatchAllTest {
+
     private StepCondition A, B, C;
     private Result result;
     private String ruleTs;
 
-    private GenericModel model;
-    private StringFieldInfo stringField;
-    private StringFieldInfo stringField2;
-    private IntegerFieldInfo zeroField;
-    private IterableFieldInfo<String, List<String>> iterableField;
-    private EnumFieldInfo<?> enumField;
+    @Test
+    void matchAll_true_true_true() throws IOException {
+        A = alwaysTrue("A");
+        B = alwaysTrue("B");
+        C = alwaysTrue("C");
+        result = when(matchAll(A, B, C)).validate().withShortCircuit(false).execute();
+        ruleTs = toTS(result.getContext().getRootMetadata());
 
-    @BeforeEach
-    void beforeEach() {
-        this.model = new GenericModel();
-        this.zeroField = model.intField(0, "zero");
-        this.stringField = model.stringField("some string", "string field 1");
-        this.stringField2 = model.stringField("other string", "string field 2");
-        this.iterableField = model.iterableField(asList("a", "b"), "list");
-        this.enumField = model.enumField(null, "enumField");
+        TypeScriptAssertionContext script = parseAs(ruleTs, TypeScriptParser::script);
+
+        assertTrue(result.value());
+        assertThat(script).numberOfSyntaxErrors().isEqualTo(0);
+        assertThat(script).identifierNamesText().containsExactly("matchAll");
+        assertThat(script).identifierReferencesText().containsExactly("DOOV");
+        assertThat(script).identifierExpressionsText().containsExactly("alwaysTrueA", "alwaysTrueB", "alwaysTrueC");
+        assertThat(script).literalsText().isEmpty();
+        assertThat(script).arrayLiteralsText().isEmpty();
     }
 
     @Test
-    void reduce_matchAll() throws IOException {
+    void matchAll_true_true_false() throws IOException {
+        A = alwaysTrue("A");
+        B = alwaysTrue("B");
+        C = alwaysFalse("C");
+        result = when(matchAll(A, B, C)).validate().withShortCircuit(false).execute();
+        ruleTs = toTS(result.getContext().getRootMetadata());
+
+        TypeScriptAssertionContext script = parseAs(ruleTs, TypeScriptParser::script);
+
+        assertFalse(result.value());
+        assertThat(script).numberOfSyntaxErrors().isEqualTo(0);
+        assertThat(script).identifierNamesText().containsExactly("matchAll");
+        assertThat(script).identifierReferencesText().containsExactly("DOOV");
+        assertThat(script).identifierExpressionsText().containsExactly("alwaysTrueA", "alwaysTrueB", "alwaysFalseC");
+        assertThat(script).literalsText().isEmpty();
+        assertThat(script).arrayLiteralsText().isEmpty();
+    }
+
+    @Test
+    void matchAll_true_false_false() throws IOException {
         A = alwaysTrue("A");
         B = alwaysFalse("B");
         C = alwaysFalse("C");
         result = when(matchAll(A, B, C)).validate().withShortCircuit(false).execute();
         ruleTs = toTS(result.getContext().getRootMetadata());
 
-        assertParenthesis(ruleTs);
         TypeScriptAssertionContext script = parseAs(ruleTs, TypeScriptParser::script);
 
         assertFalse(result.value());
@@ -91,90 +113,46 @@ class TypeScriptCombinedTest {
     }
 
     @Test
-    void reduce_and() throws IOException {
-        A = alwaysTrue("A");
+    void matchAll_false_false_false() throws IOException {
+        A = alwaysFalse("A");
         B = alwaysFalse("B");
-        result = when(A.and(B)).validate().withShortCircuit(false).execute();
+        C = alwaysFalse("C");
+        result = when(matchAll(A, B, C)).validate().withShortCircuit(false).execute();
         ruleTs = toTS(result.getContext().getRootMetadata());
-        assertParenthesis(ruleTs);
+
         TypeScriptAssertionContext script = parseAs(ruleTs, TypeScriptParser::script);
 
         assertFalse(result.value());
         assertThat(script).numberOfSyntaxErrors().isEqualTo(0);
-        assertThat(script).identifierNamesText().containsExactly("and");
-        assertThat(script).identifierReferencesText().containsExactly("alwaysTrueA");
-        assertThat(script).identifierExpressionsText().containsExactly("alwaysFalseB");
+        assertThat(script).identifierNamesText().containsExactly("matchAll");
+        assertThat(script).identifierReferencesText().containsExactly("DOOV");
+        assertThat(script).identifierExpressionsText().containsExactly("alwaysFalseA", "alwaysFalseB", "alwaysFalseC");
         assertThat(script).literalsText().isEmpty();
         assertThat(script).arrayLiteralsText().isEmpty();
     }
 
     @Test
-    void reduce_zeroInt() throws IOException {
-        result = when(zeroField.notEq(0)).validate().withShortCircuit(false).executeOn(model);
+    void matchAll_field_false_false_false_failure() throws IOException {
+        GenericModel model = new GenericModel();
+        IntegerFieldInfo zero = model.intField(0, "zero");
+        LocalDateFieldInfo yesterday = model.localDateField(LocalDate.now().minusDays(1), "yesterday");
+        StringFieldInfo something = model.stringField("something", "string field");
+        A = zero.greaterThan(4);
+        B = yesterday.after(LocalDateSuppliers.today());
+        C = something.matches("^other.*");
+        result = when(matchAll(A, B, C)).validate().withShortCircuit(false).executeOn(model);
         ruleTs = toTS(result.getContext().getRootMetadata());
-        assertParenthesis(ruleTs);
+
         TypeScriptAssertionContext script = parseAs(ruleTs, TypeScriptParser::script);
 
         assertFalse(result.value());
         assertThat(script).numberOfSyntaxErrors().isEqualTo(0);
-        assertThat(script).identifierNamesText().containsExactly("notEq");
-        assertThat(script).identifierReferencesText().containsExactly("zero");
-        assertThat(script).identifierExpressionsText().isEmpty();
-        assertThat(script).literalsText().containsExactly("0");
+        assertThat(script).identifierNamesText().containsExactly("matchAll", "greaterThan", "after", "matches");
+        assertThat(script).identifierReferencesText().containsExactly("DOOV", "zero", "yesterday", "stringfield");
+        assertThat(script).identifierExpressionsText().containsExactly("today");
+        assertThat(script).literalsText().containsExactly("4", "'^other.*'");
         assertThat(script).arrayLiteralsText().isEmpty();
     }
-
-    @Test
-    void reduce_list() throws IOException {
-        result = when(iterableField.contains("c")).validate().withShortCircuit(false).executeOn(model);
-        ruleTs = toTS(result.getContext().getRootMetadata());
-        assertParenthesis(ruleTs);
-        TypeScriptAssertionContext script = parseAs(ruleTs, TypeScriptParser::script);
-
-        assertFalse(result.value());
-        assertThat(script).numberOfSyntaxErrors().isEqualTo(0);
-        assertThat(script).identifierNamesText().containsExactly("contains");
-        assertThat(script).identifierReferencesText().containsExactly("list");
-        assertThat(script).identifierExpressionsText().isEmpty();
-        assertThat(script).literalsText().containsExactly("'c'");
-        assertThat(script).arrayLiteralsText().isEmpty();
-    }
-
-    @Test
-    void reduce_null() throws IOException {
-        result = when(enumField.isNull()).validate().withShortCircuit(false).executeOn(model);
-        ruleTs = toTS(result.getContext().getRootMetadata());
-
-        assertParenthesis(ruleTs);
-        TypeScriptAssertionContext script = parseAs(ruleTs, TypeScriptParser::script);
-
-        assertTrue(result.value());
-        assertThat(script).numberOfSyntaxErrors().isEqualTo(0);
-        assertThat(script).identifierNamesText().containsExactly("isNull");
-        assertThat(script).identifierReferencesText().containsExactly("enumField");
-        assertThat(script).identifierExpressionsText().isEmpty();
-        assertThat(script).literalsText().isEmpty();
-        assertThat(script).arrayLiteralsText().isEmpty();
-    }
-
-    @Test
-    void matches_regexp() throws IOException {
-        result = when(stringField.matches("^some.*")
-                .or(stringField2.matches("^other.*"))).validate().withShortCircuit(false).executeOn(model);
-        ruleTs = toTS(result.getContext().getRootMetadata());
-
-        assertParenthesis(ruleTs);
-        TypeScriptAssertionContext script = parseAs(ruleTs, TypeScriptParser::script);
-
-        assertTrue(result.value());
-        assertThat(script).numberOfSyntaxErrors().isEqualTo(0);
-        assertThat(script).identifierNamesText().containsExactly("matches", "or", "matches");
-        assertThat(script).identifierReferencesText().containsExactly("stringfield1", "stringfield2");
-        assertThat(script).identifierExpressionsText().isEmpty();
-        assertThat(script).literalsText().containsExactly("'^some.*'", "'^other.*'");
-        assertThat(script).arrayLiteralsText().isEmpty();
-    }
-
 
     private static String toTS(Metadata metadata) {
         final ByteArrayOutputStream ops = new ByteArrayOutputStream();
@@ -191,6 +169,7 @@ class TypeScriptCombinedTest {
         new ParseTreeWalker().walk(context, contextGetter.apply(context.getParser()));
         return context;
     }
+
     @AfterEach
     void afterEach() {
         System.out.println(ruleTs);
